% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_functions.R
\docType{package}
\name{doubleclickbidmanager_googleAuthR}
\alias{doubleclickbidmanager_googleAuthR}
\alias{doubleclickbidmanager_googleAuthR-package}
\title{DoubleClick Bid Manager API
API for viewing and managing your reports in DoubleClick Bid Manager.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:48:39
filename: /Users/mark/dev/R/autoGoogleAPI/googledoubleclickbidmanagerv1.auto/R/doubleclickbidmanager_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item 
}
}
