% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{DownloadLineItemsRequest}
\alias{DownloadLineItemsRequest}
\title{DoubleClick Bid Manager API Objects 
API for viewing and managing your reports in DoubleClick Bid Manager.}
\usage{
DownloadLineItemsRequest(fileSpec = NULL, filterIds = NULL,
  filterType = NULL, format = NULL)
}
\arguments{
\item{fileSpec}{File specification (column names, types, order) in which the line items will be returned}

\item{filterIds}{Ids of the specified filter type used to filter line items to fetch}

\item{filterType}{Filter type used to filter line items to fetch}

\item{format}{Format in which the line items will be returned}
}
\value{
DownloadLineItemsRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:11:16
filename: /Users/mark/dev/R/autoGoogleAPI/googledoubleclickbidmanagerv1.auto/R/doubleclickbidmanager_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
DownloadLineItemsRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to fetch stored line items.
}
\seealso{
Other DownloadLineItemsRequest functions: \code{\link{lineitems.downloadlineitems}}
}

