% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{DownloadLineItemsResponse}
\alias{DownloadLineItemsResponse}
\title{DownloadLineItemsResponse Object}
\usage{
DownloadLineItemsResponse(lineItems = NULL)
}
\arguments{
\item{lineItems}{Retrieved line items in CSV format}
}
\value{
DownloadLineItemsResponse object
}
\description{
DownloadLineItemsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Download line items response.
}

