% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{DownloadRequest}
\alias{DownloadRequest}
\title{DownloadRequest Object}
\usage{
DownloadRequest(fileTypes = NULL, filterIds = NULL, filterType = NULL,
  version = NULL)
}
\arguments{
\item{fileTypes}{File types that will be returned}

\item{filterIds}{The IDs of the specified filter type}

\item{filterType}{Filter type used to filter line items to fetch}

\item{version}{SDF Version (column names, types, order) in which the entities will be returned}
}
\value{
DownloadRequest object
}
\description{
DownloadRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to fetch stored insertion orders, line items, TrueView ad groups and ads.
}
\seealso{
Other DownloadRequest functions: \code{\link{sdf.download}}
}
