% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{DownloadResponse}
\alias{DownloadResponse}
\title{DownloadResponse Object}
\usage{
DownloadResponse(adGroups = NULL, ads = NULL, insertionOrders = NULL,
  lineItems = NULL)
}
\arguments{
\item{adGroups}{Retrieved ad groups in SDF format}

\item{ads}{Retrieved ads in SDF format}

\item{insertionOrders}{Retrieved insertion orders in SDF format}

\item{lineItems}{Retrieved line items in SDF format}
}
\value{
DownloadResponse object
}
\description{
DownloadResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Download response.
}
