% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{FilterPair}
\alias{FilterPair}
\title{FilterPair Object}
\usage{
FilterPair(type = NULL, value = NULL)
}
\arguments{
\item{type}{Filter type}

\item{value}{Filter value}
}
\value{
FilterPair object
}
\description{
FilterPair Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Filter used to match traffic data in your report.
}
