% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_functions.R
\name{lineitems.uploadlineitems}
\alias{lineitems.uploadlineitems}
\title{Uploads line items in CSV format.}
\usage{
lineitems.uploadlineitems(UploadLineItemsRequest)
}
\arguments{
\item{UploadLineItemsRequest}{The \link{UploadLineItemsRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/bid-manager/}{Google Documentation}

Other UploadLineItemsRequest functions: \code{\link{UploadLineItemsRequest}}
}

