% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{ListQueriesResponse}
\alias{ListQueriesResponse}
\title{ListQueriesResponse Object}
\usage{
ListQueriesResponse(queries = NULL)
}
\arguments{
\item{queries}{Retrieved queries}
}
\value{
ListQueriesResponse object
}
\description{
ListQueriesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
List queries response.
}

