% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{ListReportsResponse}
\alias{ListReportsResponse}
\title{ListReportsResponse Object}
\usage{
ListReportsResponse(reports = NULL)
}
\arguments{
\item{reports}{Retrieved reports}
}
\value{
ListReportsResponse object
}
\description{
ListReportsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
List reports response.
}
