% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{Note}
\alias{Note}
\title{Note Object}
\usage{
Note(id = NULL, message = NULL, source = NULL, timestamp = NULL,
  username = NULL)
}
\arguments{
\item{id}{Note id}

\item{message}{Message from publisher}

\item{source}{Equals 'publisher' for notification from Rubicon}

\item{timestamp}{Time when the note was added, e}

\item{username}{Publisher user name}
}
\value{
Note object
}
\description{
Note Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Publisher comment from Rubicon.
}

