% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{NotifyProposalChangeRequest}
\alias{NotifyProposalChangeRequest}
\title{NotifyProposalChangeRequest Object}
\usage{
NotifyProposalChangeRequest(action = NULL, href = NULL, id = NULL,
  notes = NULL, token = NULL)
}
\arguments{
\item{action}{Action taken by publisher}

\item{href}{URL to access proposal detail}

\item{id}{Below are contents of notification from Rubicon}

\item{notes}{Notes from publisher}

\item{token}{Deal token, available when proposal is accepted by publisher}
}
\value{
NotifyProposalChangeRequest object
}
\description{
NotifyProposalChangeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
NotifyProposalChange request.
}
\seealso{
Other NotifyProposalChangeRequest functions: \code{\link{rubicon.notifyproposalchange}}
}

