% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{Parameters}
\alias{Parameters}
\title{Parameters Object}
\usage{
Parameters(filters = NULL, groupBys = NULL, includeInviteData = NULL,
  metrics = NULL, type = NULL)
}
\arguments{
\item{filters}{Filters used to match traffic data in your report}

\item{groupBys}{Data is grouped by the filters listed in this field}

\item{includeInviteData}{Whether to include data from Invite Media}

\item{metrics}{Metrics to include as columns in your report}

\item{type}{Report type}
}
\value{
Parameters object
}
\description{
Parameters Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parameters of a query or report.
}
