% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_functions.R
\name{queries.listqueries}
\alias{queries.listqueries}
\title{Retrieves stored queries.}
\usage{
queries.listqueries()
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details. 

#' @importFrom googleAuthR gar_api_generator
}
\seealso{
\href{https://developers.google.com/bid-manager/}{Google Documentation}
}

