% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_functions.R
\name{queries.runquery}
\alias{queries.runquery}
\title{Runs a stored query to generate a report.}
\usage{
queries.runquery(RunQueryRequest, queryId)
}
\arguments{
\item{RunQueryRequest}{The \link{RunQueryRequest} object to pass to this method}

\item{queryId}{Query ID to run}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/bid-manager/}{Google Documentation}

Other RunQueryRequest functions: \code{\link{RunQueryRequest}}
}

