% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{Query}
\alias{Query}
\title{Query Object}
\usage{
Query(metadata = NULL, params = NULL, queryId = NULL,
  reportDataEndTimeMs = NULL, reportDataStartTimeMs = NULL,
  schedule = NULL, timezoneCode = NULL)
}
\arguments{
\item{metadata}{Query metadata}

\item{params}{Query parameters}

\item{queryId}{Query ID}

\item{reportDataEndTimeMs}{The ending time for the data that is shown in the report}

\item{reportDataStartTimeMs}{The starting time for the data that is shown in the report}

\item{schedule}{Information on how often and when to run a query}

\item{timezoneCode}{Canonical timezone code for report data time}
}
\value{
Query object
}
\description{
Query Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a query.
}
\seealso{
Other Query functions: \code{\link{queries.createquery}}
}

