% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{QueryMetadata}
\alias{QueryMetadata}
\title{QueryMetadata Object}
\usage{
QueryMetadata(dataRange = NULL, format = NULL,
  googleCloudStoragePathForLatestReport = NULL,
  googleDrivePathForLatestReport = NULL, latestReportRunTimeMs = NULL,
  locale = NULL, reportCount = NULL, running = NULL,
  sendNotification = NULL, shareEmailAddress = NULL, title = NULL)
}
\arguments{
\item{dataRange}{Range of report data}

\item{format}{Format of the generated report}

\item{googleCloudStoragePathForLatestReport}{The path to the location in Google Cloud Storage where the latest report is stored}

\item{googleDrivePathForLatestReport}{The path in Google Drive for the latest report}

\item{latestReportRunTimeMs}{The time when the latest report started to run}

\item{locale}{Locale of the generated reports}

\item{reportCount}{Number of reports that have been generated for the query}

\item{running}{Whether the latest report is currently running}

\item{sendNotification}{Whether to send an email notification when a report is ready}

\item{shareEmailAddress}{List of email addresses which are sent email notifications when the report is finished}

\item{title}{Query title}
}
\value{
QueryMetadata object
}
\description{
QueryMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Query metadata.
}
