% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{QuerySchedule}
\alias{QuerySchedule}
\title{QuerySchedule Object}
\usage{
QuerySchedule(endTimeMs = NULL, frequency = NULL,
  nextRunMinuteOfDay = NULL, nextRunTimezoneCode = NULL)
}
\arguments{
\item{endTimeMs}{Datetime to periodically run the query until}

\item{frequency}{How often the query is run}

\item{nextRunMinuteOfDay}{Time of day at which a new report will be generated, represented as minutes past midnight}

\item{nextRunTimezoneCode}{Canonical timezone code for report generation time}
}
\value{
QuerySchedule object
}
\description{
QuerySchedule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information on how frequently and when to run a query.
}
