% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{Report}
\alias{Report}
\title{Report Object}
\usage{
Report(key = NULL, metadata = NULL, params = NULL)
}
\arguments{
\item{key}{Key used to identify a report}

\item{metadata}{Report metadata}

\item{params}{Report parameters}
}
\value{
Report object
}
\description{
Report Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a report.
}

