% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{ReportFailure}
\alias{ReportFailure}
\title{ReportFailure Object}
\usage{
ReportFailure(errorCode = NULL)
}
\arguments{
\item{errorCode}{Error code that shows why the report was not created}
}
\value{
ReportFailure object
}
\description{
ReportFailure Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An explanation of a report failure.
}
