% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{ReportKey}
\alias{ReportKey}
\title{ReportKey Object}
\usage{
ReportKey(queryId = NULL, reportId = NULL)
}
\arguments{
\item{queryId}{Query ID}

\item{reportId}{Report ID}
}
\value{
ReportKey object
}
\description{
ReportKey Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Key used to identify a report.
}

