% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{ReportMetadata}
\alias{ReportMetadata}
\title{ReportMetadata Object}
\usage{
ReportMetadata(googleCloudStoragePath = NULL, reportDataEndTimeMs = NULL,
  reportDataStartTimeMs = NULL, status = NULL)
}
\arguments{
\item{googleCloudStoragePath}{The path to the location in Google Cloud Storage where the report is stored}

\item{reportDataEndTimeMs}{The ending time for the data that is shown in the report}

\item{reportDataStartTimeMs}{The starting time for the data that is shown in the report}

\item{status}{Report status}
}
\value{
ReportMetadata object
}
\description{
ReportMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Report metadata.
}
