% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{ReportStatus}
\alias{ReportStatus}
\title{ReportStatus Object}
\usage{
ReportStatus(failure = NULL, finishTimeMs = NULL, format = NULL,
  state = NULL)
}
\arguments{
\item{failure}{If the report failed, this records the cause}

\item{finishTimeMs}{The time when this report either completed successfully or failed}

\item{format}{The file type of the report}

\item{state}{The state of the report}
}
\value{
ReportStatus object
}
\description{
ReportStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Report status.
}

