% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{RowStatus}
\alias{RowStatus}
\title{RowStatus Object}
\usage{
RowStatus(changed = NULL, entityId = NULL, entityName = NULL,
  errors = NULL, persisted = NULL, rowNumber = NULL)
}
\arguments{
\item{changed}{Whether the stored entity is changed as a result of upload}

\item{entityId}{Entity Id}

\item{entityName}{Entity name}

\item{errors}{Reasons why the entity can't be uploaded}

\item{persisted}{Whether the entity is persisted}

\item{rowNumber}{Row number}
}
\value{
RowStatus object
}
\description{
RowStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the upload status of a row in the request.
}
