% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_functions.R
\name{rubicon.notifyproposalchange}
\alias{rubicon.notifyproposalchange}
\title{Update proposal upon actions of Rubicon publisher.}
\usage{
rubicon.notifyproposalchange(NotifyProposalChangeRequest)
}
\arguments{
\item{NotifyProposalChangeRequest}{The \link{NotifyProposalChangeRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/bid-manager/}{Google Documentation}

Other NotifyProposalChangeRequest functions: \code{\link{NotifyProposalChangeRequest}}
}

