% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{RunQueryRequest}
\alias{RunQueryRequest}
\title{RunQueryRequest Object}
\usage{
RunQueryRequest(dataRange = NULL, reportDataEndTimeMs = NULL,
  reportDataStartTimeMs = NULL, timezoneCode = NULL)
}
\arguments{
\item{dataRange}{Report data range used to generate the report}

\item{reportDataEndTimeMs}{The ending time for the data that is shown in the report}

\item{reportDataStartTimeMs}{The starting time for the data that is shown in the report}

\item{timezoneCode}{Canonical timezone code for report data time}
}
\value{
RunQueryRequest object
}
\description{
RunQueryRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to run a stored query to generate a report.
}
\seealso{
Other RunQueryRequest functions: \code{\link{queries.runquery}}
}

