% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{UploadLineItemsRequest}
\alias{UploadLineItemsRequest}
\title{UploadLineItemsRequest Object}
\usage{
UploadLineItemsRequest(dryRun = NULL, format = NULL, lineItems = NULL)
}
\arguments{
\item{dryRun}{Set to true to get upload status without actually persisting the line items}

\item{format}{Format the line items are in}

\item{lineItems}{Line items in CSV to upload}
}
\value{
UploadLineItemsRequest object
}
\description{
UploadLineItemsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to upload line items.
}
\seealso{
Other UploadLineItemsRequest functions: \code{\link{lineitems.uploadlineitems}}
}
