% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{UploadLineItemsResponse}
\alias{UploadLineItemsResponse}
\title{UploadLineItemsResponse Object}
\usage{
UploadLineItemsResponse(uploadStatus = NULL)
}
\arguments{
\item{uploadStatus}{Status of upload}
}
\value{
UploadLineItemsResponse object
}
\description{
UploadLineItemsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Upload line items response.
}
