% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleclickbidmanager_objects.R
\name{UploadStatus}
\alias{UploadStatus}
\title{UploadStatus Object}
\usage{
UploadStatus(errors = NULL, rowStatus = NULL)
}
\arguments{
\item{errors}{Reasons why upload can't be completed}

\item{rowStatus}{Per-row upload status}
}
\value{
UploadStatus object
}
\description{
UploadStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the status of upload.
}
